# docs/openapi.yaml
openapi: 3.0.3
info:
  title: IGN Shipping API
  description: Plataforma SaaS de gestão de importações
  version: 2.0.0
servers:
  - url: http://localhost:8080/api
    description: Desenvolvimento local
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
  responses:
...
    LimitExceeded:
      description: Limite do plano atingido
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    Categoria:
      type: string
      enum: [ROUPAS, TENIS, ELETRONICO, OUTROS]
    StatusPacote:
      type: string
      enum: [RASCUNHO, AGUARDANDO_ENVIO, EM_VIAGEM, ALFANDEGA, TRANSITO, ENTREGUE, FINALIZADO]
...
    StatusPagamento:
      type: string
      enum: [PENDENTE, PAGO, PARCIAL]

    # ── AUTH ─────────────────────────────────────────────────────
...
    CompradorRequest:
      type: object
      required: [nome]
      properties:
        nome:
          type: string
          minLength: 2
          maxLength: 100
        email:
          type: string
          format: email
        telefone:
          type: string
          maxLength: 20
    CompradorResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nome:
          type: string
        email:
          type: string
        telefone:
          type: string
        totalPedidos:
          type: integer
        totalGastoBrl:
          type: number
          format: double
        totalPendenteBrl:
          type: number
          format: double
        lucroGeradoBrl:
          type: number
          format: double
        codigoConvitePortal:
          type: string
          description: Código para o comprador se cadastrar no portal
    CompradorHistoricoResponse:
      type: object
      properties:
        comprador:
          $ref: '#/components/schemas/CompradorResponse'
        totalPedidos:
          type: integer
        totalPagoBrl:
          type: number
          format: double
        totalPendenteBrl:
          type: number
          format: double
        pedidos:
          type: array
          items:
            $ref: '#/components/schemas/PedidoCompradorResponse'

    # ── PRODUTO ──────────────────────────────────────────────────
...
    ProdutoResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nome:
          type: string
        categoria:
          $ref: '#/components/schemas/Categoria'
        custoYuan:
          type: number
          format: double
        freteVendedorYuan:
          type: number
          format: double
        pesoGramas:
          type: integer
        descricao:
          type: string
        fotoUrl:
          type: string
          nullable: true

    # ── PACOTE ────────────────────────────────────────────────────
...
    PedidoCompradorResponse:
      type: object
      properties:
        itemId:
          type: integer
          format: int64
        produto:
          $ref: '#/components/schemas/ProdutoResponse'
        quantidade:
          type: integer
        precoVendaBrl:
          type: number
          format: double
        statusPacote:
          $ref: '#/components/schemas/StatusPacote'
        statusPagamento:
          $ref: '#/components/schemas/StatusPagamento'
        dataPacote:
          type: string
          format: date-time
        nomeVendedor:
          type: string

    # ── RESUMO FINANCEIRO ─────────────────────────────────────────
...
    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
        message:
          type: string
        timestamp:
          type: string
          format: date-time
...
paths:
...
  /vendedor/compradores:
    get:
      tags: [Vendedor - Compradores]
      summary: Lista compradores do vendedor
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CompradorResponse'
    post:
      tags: [Vendedor - Compradores]
      summary: Cadastra novo comprador
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompradorRequest'
      responses:
        '201':
          description: Criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompradorResponse'
        '429':
          $ref: '#/components/responses/LimitExceeded'
  /vendedor/compradores/{id}:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    get:
      tags: [Vendedor - Compradores]
      summary: Detalhe do comprador
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompradorResponse'
    put:
      tags: [Vendedor - Compradores]
      summary: Atualiza comprador
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompradorRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompradorResponse'
    delete:
      tags: [Vendedor - Compradores]
      summary: Remove comprador
      responses:
        '204':
          description: Removido
  /vendedor/compradores/{id}/convite:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    post:
      tags: [Vendedor - Compradores]
      summary: Gera ou regenera código de convite para o portal
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  codigoConvite:
                    type: string
                  linkConvite:
                    type: string
  /vendedor/compradores/{id}/historico:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    get:
      tags: [Vendedor - Compradores]
      summary: Histórico de compras de um comprador
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompradorHistoricoResponse'

  # ── VENDEDOR — PRODUTOS ───────────────────────────────────────
...